[FIX] Make HgParserTest tests pass by using a rfc2822 time format
This is now correctly parsed by the hg parser.
So the hg tests now pass 100% for me.

The problem is that hg log by default ouputs a time format
that ruby's Time class does not seem to be able to handle.
example: "Tue Jan 20 11:32:54 2009 -0800"
Here the year is between time of day and timezone offset.
All of ruby's time parsing methods seem to ignore the offset
in this case.

Note that hg can be made to log in rfc2822 format by using
"hg log --template='date: {date:rfc822date}\n'", and
similarly in style files.

Michael