[FIX] BzrAdapter: rename + modify in single commit
It is possible for Bzr to report that a file has been renamed
and then modified in a single commit.

This breaks Ohloh, because a 'modify' action triggers a
comparison between the old and new files -- but since this file
has also been renamed, the old file does not exist (at least not
where Ohloh expects it).

We adapt the Bazaar system to what Ohloh expects: A rename
becomes a simple delete and add, and any subsequent modification
is ignored.