Specify module name when using CVS test repository
During production operation, we always force users to specify
a module name. During test our tests, we are using an empty string
as the module name.

CVSNT has varying behavior when the module name is left empty.
In some builds (Mac development machines), leaving out the module
name causes rlog to include the CVSROOT directory, while in other
builds (production machines) the CVSROOT directory is excluded.

The presence or absence of the CVSROOT directory causes a unit test
to pass or fail with different cvsnt builds.

The flaw here is in the test harness: it should not leave the module
name empty, and thus the CVSROOT directory should never be logged.

I have modified the test harness to specify the CVS module name
everywhere. I have modified the expected responses from CVS to
exclude the CVSROOT directory.