OTWO-443: Fixes GitStyledParser to support empty commits
If a commit did not contain any changes at all, then the GitStyledParser
would fail to parse the commit.

When changes are present, `git whatchanged` inserts an extra newline.
When changes are not present, we don't get the changes listed, and we
also don't get the extra newline.

That newline was required to trigger the parser to advance states.

I've added an explicit newline in our log format string to guarantee
that we'll always get at least one newline. If we get two, that's OK.

At some point we should probably get rid of GitStyledParser completely,
and use something more reliable like `git cat-file` or `git
fast-export`.