OTWO-623 Fixes incorrect list of diffs for null git merges
Whenever a git merge commit has a resulting tree identical to its
first parent's tree, the list of diffs should be empty.

However, a bug in our commit parser caused us to return the set of
diffs relative to the second parent, instead of the correct empty list.

I could not invent a method to force `git whatchanged` to do what we
want, so I surrendered, and instead I manually patch the result after
the fact.