OTWO-208 Ignore 'fatal signal 11' from all CVS servers
We have always ignored 'fatal signal 11' when it is returned from
eclipse.org in response to a `cvs ls` command. We do this because
eclipse.org *always* returns this error, even when everything is OK.

As it turns out, more CVS servers than just eclipse.org do this. There
is probably a particular version or configuration of CVS that causes
this.

The result is that many CVS repositories cannot be processed by Ohloh
because we treat these errors as fatal.

Let's simply accept 'fatal signal 11' as standard operating procedure,
and always ignore it when it is returned by `cvs ls`.