OTWO-840 Removes redundant log query from BzrAdapter.cat_file_parent
This patch allows BzrAdapter to process repositories that include ghost
commits, and should also improve line count performance considerably.

BzrAdapter's cat_file_parent() was needlessly complicated. To determine
the revision ID of a parent commit, we would request the log for that
commit -- using the revision ID! -- and then parse the revision ID from
the resulting log fragment. This work was completely pointless.

Worse, in the case of a ghost commit, the parent commit is not actually
present in the repository. This would cause the log request to fail, and
thus cat_file_parent() would fail.

Ohloh doesn't actually care what the ghost commit contains; we were
trying to read it simply to find its ID (which we already know, and
worse, won't use). By removing the redundant ID lookup, Ohloh can
successfully process repositories that contain ghost commits.

I spent some time trying to create a test repostiory that includes a
ghost commit, but this is rather involved. You can't do it using the bzr
CLI alone. The most common source of ghost commits seems to be some
complex interaction with bzr svn-import.

I manually verified that this fix works with repository `lp:bzr` itself,
which has a ghost commit as a parent of revo 897.