Add string encoder to replace invalid utf8 chars
Iconv is the standard bash tool for handling encodings. We tried to
replace our ruby string_encoder with it. We were able to convert invalid
characters, however we couldn't get the output string length to match
the input string length.

$ cat spec/data/invalid-utf-word | iconv -f UTF-8 -t ASCII//TRANSLIT

We need more time to explore this. For now we are moving the
string_encoder as it was. We have string_encoder as a separate binary
because we believe this will reduce the memory footprint. We currently
pipe scm logs into string_encoder in a subprocess. If we were to read
that log into the main ruby process and then fix the encoding, it will
add to the memory of the main process. We believe that the memory would
free up sooner if we run string_encoder operation in a subprocess.