exist? must rely on raising exceptions like in v2
The exist? method in version 2 relied upon purposefully raising
exceptions to indicate that a repository is missing. In v3 we tried to
replace this with a check for scm_dir(.git, .hg etc) instead. However we
now realize that exist? needs also to work for remote urls. So we are
reverting back to the ohloh_scm v2 logic:

https://github.com/blackducksoftware/ohloh_scm/blob/v2.5.1/lib/ohloh_scm/adapters/git/misc.rb#L7